Chore: Made webpage mobile responsive
