+++ frontend/src/app/layout.tsx
@@ -1,27 +1,27 @@
-import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
-import './globals.css'
-import { StarknetProvider } from './components/StarknetProvider'
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import "./globals.css";
+import { StarknetProvider } from "./components/StarknetProvider";
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: 'Starknet Scaffold',
-  description: 'Created with love',
-}
+  title: "Starknet Scaffold",
+  description: "Created with love",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
-    <html lang="en" className='dark'>
-      <body className={`${inter.className} dark:bg-black bg-gray-300 dark:text-white`} >
-        <StarknetProvider>
-          {children}
-        </StarknetProvider>
+    <html lang="en" className="dark">
+      <body
+        className={`${inter.className} dark:bg-black bg-gray-300 dark:text-white transition-all duration-500 ease-in-out`}
+      >
+        <StarknetProvider>{children}</StarknetProvider>
       </body>
     </html>
-  )
+  );
 }
